requirements.txt
langgraph>=0.2.0
langchain-google-genai>=1.0.0
google-cloud-bigquery>=3.13.0
google-cloud-bigquery-storage>=2.22.0
pyarrow>=14.0.0
pandas>=2.0.0
python-dotenv>=1.0.0 
langchain_core>=0.3.0
db-dtypes==1.2.0
pytest>=7.4.0
pytest-mock>=3.11.0